Changed type of status

+++ contracts/IDisputeResolutionLayer.sol
@@ -2,6 +2,6 @@ pragma solidity ^0.4.18;
 
 //TODO: Should maybe enforce checkProofOrdered
 interface IDisputeResolutionLayer {
-    function status(bytes32 id) public view returns (uint8); //returns State enum
+    function status(bytes32 id) public view returns (uint); //returns State enum
     function commitChallenge(address solver, address verifier) public returns (bytes32 gameId);
 }
